feat: 支持设置 video 和 page 的下载并发
